refactor: setting 관련 파일을 feature/setting 폴더 내로 이동
