+++ lib/features/auth/data/data_sources/remote_kakao_auth_data_source.dart
@@ -4,8 +4,8 @@ import 'package:dio/dio.dart';
 import 'package:flutter/services.dart';
 import 'package:mongbi_app/core/exceptions/auth_custom_exception.dart';
 import 'package:mongbi_app/core/secure_storage_service.dart';
-import 'package:mongbi_app/data/dtos/user_dto.dart';
 import 'package:mongbi_app/features/auth/data/dtos/login_response_dto.dart';
+import 'package:mongbi_app/features/setting/data/dtos/user_dto.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 
 class RemoteKakaoAuthDataSource {
