+++ lib/features/auth/data/dtos/login_response_dto.dart
@@ -1,4 +1,4 @@
-import 'package:mongbi_app/data/dtos/user_dto.dart';
+import 'package:mongbi_app/features/setting/data/dtos/user_dto.dart';
 
 class LoginResponseDto {
   LoginResponseDto({required this.token, required this.user});
