+++ lib/features/setting/data/data_sources/remote_user_info_data_source.dart
@@ -1,7 +1,7 @@
 import 'package:dio/dio.dart';
 import 'package:mongbi_app/core/secure_storage_service.dart';
-import 'package:mongbi_app/data/data_sources/user_info_data_source.dart';
-import 'package:mongbi_app/data/dtos/user_dto.dart';
+import 'package:mongbi_app/features/setting/data/data_sources/user_info_data_source.dart';
+import 'package:mongbi_app/features/setting/data/dtos/user_dto.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 
 class RemoteUserInfoGetDataSource implements GetUserInfoDataSource {
