+++ lib/features/setting/data/repositoreis/remote_user_info_reposiotory.dart
@@ -1,6 +1,6 @@
-import 'package:mongbi_app/data/data_sources/user_info_data_source.dart';
-import 'package:mongbi_app/domain/repositories/user_info_repository.dart';
 import 'package:mongbi_app/features/auth/domain/entities/user.dart';
+import 'package:mongbi_app/features/setting/data/data_sources/user_info_data_source.dart';
+import 'package:mongbi_app/features/setting/domain/repositories/user_info_repository.dart';
 
 class RemoteUserInfoRepository implements UserInfoRepository {
   final GetUserInfoDataSource dataSource;
