+++ lib/features/setting/presentation/models/bgm_state.dart
