+++ lib/features/setting/presentation/pages/widgets/nickname_title.dart
