+++ lib/features/setting/presentation/view_models/nickname_setting_view_model.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:mongbi_app/features/auth/domain/entities/user.dart';
-import 'package:mongbi_app/providers/nickname_provider.dart';
-import 'package:mongbi_app/providers/user_info_provider.dart';
+import 'package:mongbi_app/features/setting/providers/nickname_provider.dart';
+import 'package:mongbi_app/features/setting/providers/user_info_provider.dart';
 
 class NicknameViewModel extends Notifier<User?> {
   @override
