+++ lib/features/setting/providers/nickname_provider.dart
@@ -1,11 +1,11 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:mongbi_app/data/data_sources/nickname_setting_data_source.dart';
-import 'package:mongbi_app/data/data_sources/remote_nickname_setting_data_source.dart';
-import 'package:mongbi_app/data/repositories/remote_nickname_setting_repository.dart';
 import 'package:mongbi_app/domain/repositories/nickname_repository.dart';
-import 'package:mongbi_app/domain/use_cases/nickname_setting_use_case.dart';
 import 'package:mongbi_app/features/auth/domain/entities/user.dart';
-import 'package:mongbi_app/presentation/setting/view_models/nickname_setting_view_model.dart';
+import 'package:mongbi_app/features/setting/data/data_sources/nickname_setting_data_source.dart';
+import 'package:mongbi_app/features/setting/data/data_sources/remote_nickname_setting_data_source.dart';
+import 'package:mongbi_app/features/setting/data/repositoreis/remote_nickname_setting_repository.dart';
+import 'package:mongbi_app/features/setting/domain/use_cases/nickname_setting_use_case.dart';
+import 'package:mongbi_app/features/setting/presentation/view_models/nickname_setting_view_model.dart';
 import 'package:mongbi_app/providers/core_providers.dart';
 
 final _nicknameSettingDataSourceProvider = Provider<NicknameSettingDataSource>(
