+++ lib/features/setting/providers/user_info_provider.dart
@@ -1,10 +1,10 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:mongbi_app/data/data_sources/remote_user_info_data_source.dart';
-import 'package:mongbi_app/data/repositories/remote_user_info_reposiotory.dart';
-import 'package:mongbi_app/domain/repositories/user_info_repository.dart';
 import 'package:mongbi_app/features/auth/domain/entities/user.dart';
 import 'package:mongbi_app/features/auth/domain/use_cases/get_user_info_use_case.dart';
 import 'package:mongbi_app/features/auth/presentation/view_models/auth_view_model.dart';
+import 'package:mongbi_app/features/setting/data/data_sources/remote_user_info_data_source.dart';
+import 'package:mongbi_app/features/setting/data/repositoreis/remote_user_info_reposiotory.dart';
+import 'package:mongbi_app/features/setting/domain/repositories/user_info_repository.dart';
 import 'package:mongbi_app/presentation/splash/view_models/splash_state.dart';
 import 'package:mongbi_app/presentation/splash/view_models/splash_view_model.dart';
 import 'package:mongbi_app/providers/core_providers.dart';
